# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Users System Api  Documentation ",
        "description": "This documentation for user system this system hav a Users And Admins And Moderators And Posts ",
        "contact": {
            "email": "abdulrahman.masoud.official@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "This Api For User Project"
        }
    ],
    "paths": {
        "api/user": {},
        "/api/user/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Posts User System",
                "description": "Posts In User System",
                "operationId": "App\\Http\\Controllers\\PostController::index",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login In User System",
                "description": "Login In User System If U r Admin orn Mod Or User",
                "operationId": "userLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Sender Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Sender password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "ٌRegister In User System",
                "description": "ٌRegister In User System",
                "operationId": "userRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "you must add Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "you must add Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "you must add password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/user/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "ٌLogoutIn User System",
                "description": "ٌLogout In User System",
                "operationId": "userLogout",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "API Endpoints of User Project"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}